+++ src/motion.h
@@ -3,9 +3,9 @@ class Motion {
         Motion();
         void move(int angle, double speed, int rot);
         void panic();
-        int fast_sin(int angle);
-        int fast_cos(int angle);
+        float fast_sin(int angle);
+        float fast_cos(int angle);
     private:
         int _pins[4][2];
-        int _sin[91];
+        float _sin[91];
 };
